draw_graph() is moved to oemofm/, and unnecessary imports are removed
