+++ mimo.py
@@ -19,80 +19,11 @@ import connection_oep as conn
 
 # misc.
 import os
-import numpy as np
-import pandas as pd
-import networkx as nx
-import matplotlib.pyplot as plt
-from datetime import datetime
 
 
 ###############################################################################
-# Helper Functions
+# Comparison Function
 ###############################################################################
-
-
-def draw_graph(grph, edge_labels=True, node_color='#AFAFAF',
-               edge_color='#CFCFCF', plot=True, node_size=2000,
-               with_labels=True, arrows=True, layout='neato'):
-    """
-    Draw a graph. This function will be removed in future versions.
-
-    Parameters
-    ----------
-    grph : networkxGraph
-        A graph to draw.
-    edge_labels : boolean
-        Use nominal values of flow as edge label
-    node_color : dict or string
-        Hex color code oder matplotlib color for each node. If string, all
-        colors are the same.
-
-    edge_color : string
-        Hex color code oder matplotlib color for edge color.
-
-    plot : boolean
-        Show matplotlib plot.
-
-    node_size : integer
-        Size of nodes.
-
-    with_labels : boolean
-        Draw node labels.
-
-    arrows : boolean
-        Draw arrows on directed edges. Works only if an optimization_model has
-        been passed.
-    layout : string
-        networkx graph layout, one of: neato, dot, twopi, circo, fdp, sfdp.
-    """
-    if type(node_color) is dict:
-        node_color = [node_color.get(g, '#AFAFAF') for g in grph.nodes()]
-
-    # set drawing options
-    options = {
-     'prog': 'dot',
-     'with_labels': with_labels,
-     'node_color': node_color,
-     'edge_color': edge_color,
-     'node_size': node_size,
-     'arrows': arrows
-    }
-
-    # draw graph
-    pos = nx.drawing.nx_agraph.graphviz_layout(grph, prog=layout)
-
-    nx.draw(grph, pos=pos, **options)
-
-    # add edge labels for all edges
-    if edge_labels is True and plt:
-        labels = nx.get_edge_attributes(grph, 'weight')
-        nx.draw_networkx_edge_labels(grph, pos=pos, edge_labels=labels)
-
-    # show output
-    if plot is True:
-        plt.show()
-
-
 def comparison(u_model, o_model, threshold=0.1):
     """
     Function for comparing urbs & oemof
@@ -219,10 +150,10 @@ def create_om(input_data, timesteps):
     graph = False
     if graph:
         graph = create_nx_graph(es, model)
-        draw_graph(graph, plot=True, layout='neato', node_size=3000,
-                   node_color={'b_0': '#cd3333',
-                               'b_1': '#7EC0EE',
-                               'b_2': '#eeac7e'})
+        oemofm.draw_graph(graph, plot=True, layout='neato', node_size=3000,
+                          node_color={'b_0': '#cd3333',
+                                      'b_1': '#7EC0EE',
+                                      'b_2': '#eeac7e'})
 
     # get results
     es.results['main'] = outputlib.processing.results(model)
