+++ oemofm/__init__.py
@@ -2,4 +2,4 @@
 
 """
 
-from .model import create_model
+from .model import *
